Improve fidelity
Some grammar rules improperly consume EOL (with a `\\n` or `^` match)
which prevents the _following_ line from getting formatted correctly.
In the process of fixing those (with `(?=\\n)`, a non-consuming EOL
lookahead), I've made some additional enhancements.

Add #line-continuation to make reasoning about that easier and more
consistent in the rules.

Add #rule to separate that logic from #recipe.

Add support for rules with inline recipes (';'-prefixed).

Fix #comment to *not* consume trailing newline, because that borks up
any containing grammar (a comment ending a rule line used to result in
the first recipe line *not* getting colorized!).

Rename #directives and #variables to #directive and #variable for
consistency (the grammar handles an individual directive or variable,
not a set of them).

Change directives to allow leading whitespace, since make seems to allow
it.

Remove delegation to "source.string", because Makefiles need to show
*make's* handling of variable interpolation more than the shell's.
Removed "interpolation" string (backticks) for the same reason.

Change several 'end' regexps that consumed end-of-line ("\\n" or "^")
and replaced them with look-aheads ("(?=\\n)") to fix cases where the
*following* line did not get matched correctly.